Fix show timezones again
