+++ src/components/Timeline/TimelineDate.astro
@@ -19,6 +19,7 @@ const now = new Date();
           year: "numeric",
           hour: "numeric",
           minute: "2-digit",
+          timeZone: "Europe/Oslo",
         })
       }
     </span>
