+++ src/pages/om_oss/index.astro
@@ -13,7 +13,7 @@ import TimelineDate from "@/components/Timeline/TimelineDate.astro";
 import ProfilePicture from "@/components/Images/ProfilePicture.astro";
 import { Image } from "astro:assets";
 
-import { SHOWS } from "@/lib/constants";
+import { SHOWS } from "@/lib/shows";
 import { getEntry } from "astro:content";
 import { getCollection } from "astro:content";
